extract initial functions to init file
